fixed H-01, M-01, L-02, CON-1, PFA-1, POO-1, SPR-1

+++ contracts/signed/SignedSpotPricer.sol
@@ -23,7 +23,7 @@ contract SignedSpotPricer is SpotPricer, OwnableUpgradeable {
     emit SettlePrice(expiry, price, 0);
   }
 
-  function checkRoundId(uint /* expiry */, uint /* _roundId */) internal pure override returns (bool) {
+  function checkRoundId(uint /* expiry */, uint80 /* _roundId */) internal pure override returns (bool) {
     return false;
   }
 }
